First Version of Zeppelin

+++ src/main/java/lielietea/mirai/plugin/core/game/zeppelin/data/Coordinate.java
@@ -1,15 +1,15 @@
 package lielietea.mirai.plugin.core.game.zeppelin.data;
 
 public class Coordinate {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Coordinate(int x,int y){
-        this.x=x;
-        this.y=y;
+    public Coordinate(double x, double y) {
+        this.x = x;
+        this.y = y;
     }
 
-    public static boolean equals(Coordinate c1, Coordinate c2){
-        return (c1.x==c2.x && c1.y==c2.y);
+    public static boolean equals(Coordinate c1, Coordinate c2) {
+        return (c1.x == c2.x && c1.y == c2.y);
     }
 }
